Implemented OnLoad for Button.
The Json structure for a button requires that the label be defined as its own object
with the text field defined. In the future, the type should be changeable to other
controls such as images.

The change also required moving the logic to update the button size into its own
function.