Window::Load function for a Json object.
The Json object allows for a window's appearance to be defined such as the title, size,
and what controls should be added to the body of the window.

Application now has a NewWindow function for a Json string
stream.

The function will parse the Json string into a Json object and pass it to the updated
CreateWindow function.

The CreateWindow private function now uses a Json object to define the parameters
for the window, which is passed into the window's Load function.