Overview app uses Json streams to define UI controls.
The controls for the main window are defined in an Overview.json file. This file is
under the Resources folder and is copied to the output binary directory as the
post build step.

This is all temporary and the proper method of loading these json streams will be
defined as the library gets fleshed out.