Initial implementation of ControlList class.
This class will help with gathering controls with a defined ID. If IDs have been defined
in a json file, loading the file with the this object will allow setting up callbacks and
other custom logic for the desired control.

There is currently no safe mechanism in place if trying to access a control with an
invalid ID. The program will crash trying to operate on an invalid pointer. This should
be resolved in a future commit.