Refactored clip properties to its own class and implemented clip offset.
The clip offset is used to move the clipping rectangle to view different parts of the
clipped region.

The first implementation to use this feature is the TextInput control. When the cursor
is moved out of the view area, the clipping offset will be updated to show the area
around the cursor.