ScrollableContainer+ScrollBar: Finished implementation.
The ScrollableContainer now scrolls properly with the full contents of the container
based on the normalized position of the scroll bar handle. The vertical scroll bar
will also be smaller if the horizontal scrollbar exists.

ScrollBar has also been refactored to only manage the control itself, and not keep
any information about the size of the contents it may be manipulating.

ScrollBar sizes have been added to the Theme.