Text: Improved positioning.
The text control's size was variable due to the size of the text contents. This
could vary for each text because of the differences in descent for characters.
The control will now use the font size as the total height, and will center
the text within that size based on the content size of the text.

This change allows us to make improvements to the Font's API as well as fixing
an issue where the max height wasn't being used when measuring text.