Application: Ability to initialize with json stream.
A json stream can now be used to define the windows used
in an application as well as the font and font size to use. The
function takes in a map of a string and ControlList object
that allows the caller to find a window and set up callbacks
to controls that have been given an ID.

The json stream must contain a 'Main' window definition, and
all windows must have a defined ID field set.