ListBox: Add control with basic functionality.
This control can contain a number of controls in a list that
is scrollable. The contents are also clipped to ensure that
only the visible region is shown.

The Json loading path will forward the 'Controls' declaration
to the proper container while all other properties only affect
the ListBox control.