Interface: Changed the way SFML interface clips the view.
Previously, the view was clipped by rendering to a backing texture,
then was converted to a sprite, then clipped and placed into the
render target.

The new method is to use the sf::View functionality to restrict the
view to what the clip bounds are.

This is a much faster operation and the improvements are easily
noticeable on a weak linux machine, which the tests were performed
on.