Text: Store shared_ptr to font object.
This will point to the default font in the loaded theme object.
All queries about font information is converted to ask the text
object if it has one. This allows for text controls to have a
different font loaded from the default font.

The Paint functions have been modified to now take in a font
object.