CommandPalette: Tool for commands to be executed.
A new Tools sub-folder has been created to place any controls that can be used
to help with debugging. Controls in this folder will only be compiled if
DTTOOLS=ON is passed into the CMake generator. These files are not compiled
in by default.

The main window will create the CommandPalette control on init when TOOLS is
defined. This mimics the behavior found in other text editors with Ctrl+P key
combination to bring up a text input field to issue commands from.