Tree: Add plus/minus icons for classic theme.
The classic theme will use plus/minus icons for the tree toggle image. A box is
also painted around this image. Selection painting has been updated to only
paint around the text control of the tree item.

The toggle image color will match the color of the text control if the whole row
is painted when selected.