Inspector: Initial commit of the Inspector window.
When typing the command 'Inspector' into the command palette, a new window
will be created to allow inspecting the window from which the command
originated from. This window will display the full list of controls in a tree view.

This is the first step. More information will be added to allow inspecting details
of each control.