Properties: Initial commit of a properties control.
This control can take a Json object and do a shallow parse of the object and
add each item as text to the list when a tree is selected.

The Inspector will attach the associated control in the target window to the
tree so that it's properties can be inspected.