Profiler: Event type to hold elapsed and count data.
The Sample class is now mainly used for adding an event when it goes out
of scope.

The Frame class now holds a list of events. It also now has its own sample object
which will manually call BeginSample/EndSample when it goes out of scope.
The final event of a frame corresponds to the total frame time.

The Profiler header has been fixed to properly create an object on the stack.