Profiler: Add sample groups to build a call tree.
A new macro is introduced, PROFILER_SAMPLE_GROUP, which is another sample
but is the start of another tree of events that can be sampled.

The Frame object now contains a single event record as the root and all other
samples will be added to this root.

The ProfileViewer is updated to show any child events within any groups.