TextureViewer: Tool to help view currently loaded textures.
This tool can be opened through the command palette with the "textureviewer"
or "tv" command.

When opened, the tool will find all loaded textures in the TextureCache object. Then
all loaded fonts are added and then the Icons.

A list is shown for these textures and can be selected to view the contents. The size
along with the texture is displayed.