Font: Rewrite font loading to support multiple character ranges.
A character range is defined as min and max integer values. These font ranges
are then used to load all glyphs for each range. Multiple ranges can be requested
to be loaded. The logic has also been improved to attempt to all of the bitmap
data into the smallest possible texture size by incrementing by 128 for width then
height.

Glyphs are not stored in the font as an unorder map between the integer value
and the UV coordinates into the texture. If a given codepoint is not found, then a
missing character glyph bitmap is used in its place.

Calls to draw and measure strings have been updated to not need to apply a -32
offset due to the conversion to an unordered map.