Font: Update Draw function to use the packed quad data.
The previous way of drawing characters was based on the stbtt_GetBakedQuad
function. The update uses the stbtt_GetPackedQuad function as a template, but is
custom to this library.

All Font APIs have been changed to operate on uint32_t types.

Also updates all loops to use char32_t instead of char.