MenuBar: Keep MenuItem selected when open.
The opened menu item was not remaining selected when the
mouse was not hovering over the item. The opened menu
item should continue to have the hovered state look even
if the mouse is not hovered to indicate which menu item is
being viewed.

This change required that the OnTextSelectableSignature
callback use a non-const parameter.