ScrollableViewControl: Add m_IgnoreOwnedControls flag.
This flag allows the GetControl function to skip any controls that this parent
container owns. This is to handle cases where other controls like a Panel to be added
and ignored so that the Interaction is returned when necessary.

This commit also adds the interaction object to the list of controls, but still
performs a Contains check of the container before returning the interaction.

Added getters for the interaction as well as the owning ScrollableViewControl
inside of ScrollabeViewInteraction.