Control: Change OnInvalidate to accept a shared ptr.
The OnInvalidate callback now uses a shared ptr instead of a
raw pointer. This was due to the control itself invalidating and
not having the functionality to generate a shared ptr. Since
Controls now inherit from std::enable_shared_from_this, this
is now possible. Outside objects can now store a shared pointer
or a weak pointer to operate on a possibly valid/invalid Control
at a later time.

The big use case for this is the Window object. This object
captures all containers that have been invalidated and attempts
to call the Layout function on them. This is now memory safe
as previously it was operating on raw pointers.