Frontend: Remove direct path references for SFML.
This is the start of a rewrite of how the Frontend library is structured. Previously,
the library dependencies were hardcoding paths to the libraries to link. This commit
along with future commits will attempt to move away from this method and use
the configuration files associated with the dependency. The developer's build system
should set up the paths to these dependencies.

This commit address SFML on the Windows platform.