Frontend: Rework code and structure for more flexibility.
The Frontend library has now been split into two modules, the Rendering and
Windowing modules. This will allow frontends to mix and match a windowing
system with a separate rendering system.

This commit rewrites the SFML frontend into its own windowing and
rendering modules. The default on Windows and Linux is to use SFML as
both the windowing and rendering modules.

The Interface files have been deprecated in favor of this new method.