Application: Intialize function has been rewritten.
This rewrite makes it so that each defined window is created and
initialized with its title and size settings. Then the 'Main' window
is displayed and created from the 'Windowing' module such as
SDL2/SFML/etc. Then the contents of the window like the menubar
and body controls are loaded last.

The reason for this is that some rendering platforms require a window
to be created before any textures/rendering calls can be made.