Json: Add 'IsError' flag.
This flag allows the parser to propagate any errors that may occur during parsing
of a Json string. An error handling function has been added that will create a
Json error object with the 'Line', 'Column', and 'Error' keys.

Error conditions and messages have been added for parsing aspects of an Object.