Paint: Add Arc and RectangleRounded functions.
The Arc function will draw a quarter of a circle and an angle offset can be
given to move the arc.

The RectangleRounded function will paint a rectangle with rounded corners.

The Circle function has been fixed to only add each generated vertex once and
properly index them. Previously, copies of each vertex were being added to
match the indices.