Application: OnWindowAction callback.
Create/Destroy functionality of windows has been moved to an
enum WindowAction. This will allow for easier extensibility when
adding new actions such as raising/minimizing/maximizing etc.
By having this as an enum, the OnCreate/OnDestroy callbacks can
be deprecated and all actions can be handled in a single
callback.