Application: 'Raise' window action.
This action will notify the Frontend that a particular window
will need to be raised. The first use for this are modal windows.
These windows should always be on top.

The 'About' window in the 'Overview' app is set as a modal window.

The 'SDL2' frontend has implemented the 'Raise' window action.