Application: Provide API function 'OpenFileDialog'.
This provides a single entry point for opening the file dialog. The goal is
to allow either the system file dialog to be open or a custom one. This is due
to some platforms not providing an easy access to one such as Linux, but also
allows for the custom one to be shown on every platform for uniformity.