Frontend: Register NewFrame callback.
The NewFrame callback allows for the Windowing/Rendering system
to perform and per-frame initialization/cleanup. The first instance
of this is in SDL2, where unhandled window events can occur for
a particular window. These events are pushed into a vector and
processed when the desired window is being asking for the next
event to process. On each new frame, this vector is cleared to
prevent any buffer accumulation that may occur.

One major reason for this change is that SDL2 can report an event
where there is no associated window i.e. windowID of 0. If this is
the case, then all windows should handle the event.