Windowing: Robust event handling in SDL2.
All windows will attempt to pump events from the SDL event queue.
If any events come in for a window that is not the current, then
the event is placed into a queue for later processing when that
window begins its event pumping.

On some platforms such as Linux, a window ID of '0' is returned.
For these cases, the event should be handled by all windows. An
example of this is the mouse up event when no window is hovered.