Texture: API functions for loading SVG files.
SVG files are loaded using the NanoSVG library. The API allows for loading the
SVF file with the width and height defined by the file, or it can be defined
through the API function. This will allow for scaling the SVG image to the
requested size.

The 'Load' function has also been refactored to load either an SVG or PNG file
based on the extension.