Application: Json property to load a set of images per icon type.
A new Json property 'Icons' is introduced. This object contains a 'Size' property
that defines the size of each icon, a 'Types' array to define the type of icon
and SVG file associated with the icon, and a 'UseInternal' flag. This is used to
revert the icons to use an internal bitmap defined in the code.