Windowing: Handle WindowAction::Position.
When the frontend receives this action, the windowing subsystem should be
notified and update the corresponding window accordingly.

To properly handle smooth movement of the window, a hit test callback needs
to be registered. The callback needs to return whether the point given is
within a draggable area such as the title bar or the normal area such as the
contents of the window. SDL already provides an API for this while SFML
does not include one. To support SFML, a custom hit test callback is
introduced in the Windows subsystem.