Class: Split name into name and full_name.
The full_name is the fully qualified name of the class which
includes any namespaces the class is a part of. The class member
is now just the class name without any namespaces.

The set_full_name function will attempt to split the full_name
into separate fields and extract just the class name if possible.