Manifest: Generate .md file for each class.
A new generate function is introduced which will attempt to
create markdown files for each class found when parsing the doxygen
xml.

A new output-dir argument can be accepted which specifies the root
directory at where the 'md' output directory will be created. This
is the location where the Manifest will place all generated
markdown files.