Tree: Use direct UVs width and text height to determine list offset.
Previously, the offset was determined by the position and size of the text
control. This means the offset won't be computed until a layout occurs. This
will not take effect if the tree is initialized upon creation. This means the list
won't be properly positioned until it is collapsed/expanded again.