Splitter: Moved Resize call to PlaceControls.
Previously, the Resize function was called in OnLayoutComplete. Due to a change
that is being made with how the Window object handles invalidations, this
needed to be moved into the PlaceControls  function so that an extra layout
request is not needed.

The Table also needed to initialize it's container's size when creating a column
so that the contents can be properly laid out.