TextInput: Fix issue with adding newline at the beginning with existing text.
This issue can be reproduced with a TextInput control that contains text with
multiple lines. When a newline character is added at the beginning of this
text buffer, the cursor would be properly placed but new text added would be
placed in the incorrect position.

The fix is to make sure that the 'LineStartIndex' function doesn't wrap when
subtracting from an index that is already at the beginning of the text buffer.