Texture: Fix issue with passing invalid loaded width and height from SVG.
There was an issue where the width and height of a loaded SVG image was
being passed to the 'Load' function after rasterizing the image. Unfortunately,
this function also destroys the loaded image and points to an invalid width
and height. This is now fixed by storing these values before rasterizing.