Tools: Introduction of 'Commands' object.
This object is responsible for holding a list of registered commands and invokes a command given the name and a set of arguments.

The CommandPalette now invokes a callback if a given command
is not found. The Tools interface will pass the information
provided in the callback to the 'Commands' object and will
invoke the proper command if it is registered.