Splitter: Fix mouse cursor issue.
When the mouse is hovering over a splitter, the cursor changes to the
appropriate icon based on the splitter orientation. There was an issue
where once the Splitter control gains focus, moving the mouse off of
the splitter would sometimes result in the cursor not reverting back
to the arrow. This commit fixes that issues and also makes sure it
does not override mouse cursor changes from other controls.